requirements.txt
Flask ~= 2.3.3
beautifulsoup4 ~= 4.12.2
html5lib ~= 1.1
requests ~= 2.31.0
python-dotenv ~= 1.0.0
flask-cors ~= 4.0.0
pybase62 ~= 1.0.0